+++ src/app/config/app-config.ts
@@ -2,9 +2,9 @@ import * as OJP from '../shared/ojp-sdk/index'
 
 export interface AppMapLayerOptions {
     LIR_Restriction_Type: OJP.GeoRestrictionType
-    LIR_POI_Type?: null | OJP.GeoRestrictionPoiOSMTag
+    LIR_POI_Type?: null | OJP.GeoRestrictionPoiOSMTag | OJP.GeoRestrictionPoiOSMTag[]
     minZoom: number
-    layer_ids: string[],
+    layer_ids?: string[] | null,
     click_layer_ids?: string[] | 'SAME_AS_LAYER_IDS' | null
 }
 
@@ -45,7 +45,7 @@ const map_app_map_layers: Record<string, AppMapLayerOptions> = {
         click_layer_ids: 'SAME_AS_LAYER_IDS',
     },
     'sharing_cars': {
-        LIR_Restriction_Type: 'poi',
+        LIR_Restriction_Type: 'poi_amenity',
         LIR_POI_Type: 'car_sharing',
         minZoom: 12,
         layer_ids: [
@@ -54,7 +54,7 @@ const map_app_map_layers: Record<string, AppMapLayerOptions> = {
         click_layer_ids: 'SAME_AS_LAYER_IDS',
     },
     'sharing_bicycles': {
-        LIR_Restriction_Type: 'poi',
+        LIR_Restriction_Type: 'poi_amenity',
         LIR_POI_Type: 'bicycle_rental',
         minZoom: 14,
         layer_ids: [
@@ -63,7 +63,7 @@ const map_app_map_layers: Record<string, AppMapLayerOptions> = {
         click_layer_ids: 'SAME_AS_LAYER_IDS',
     },
     'sharing_scooters': {
-        LIR_Restriction_Type: 'poi',
+        LIR_Restriction_Type: 'poi_amenity',
         LIR_POI_Type: 'escooter_rental',
         minZoom: 14,
         layer_ids: [
@@ -72,11 +72,21 @@ const map_app_map_layers: Record<string, AppMapLayerOptions> = {
         click_layer_ids: 'SAME_AS_LAYER_IDS',
     },
     'charging_stations': {
-        LIR_Restriction_Type: 'poi',
+        LIR_Restriction_Type: 'poi_amenity',
         LIR_POI_Type: 'charging_station',
         minZoom: 12,
         layer_ids: [
             'charging-station-icon',
+        ],
+        click_layer_ids: 'SAME_AS_LAYER_IDS',
+    },
+
+    'pois-ALL': {
+        LIR_Restriction_Type: 'poi_all',
+        LIR_POI_Type: ['service', 'shopping', 'leisure', 'catering', 'public', 'parkride', 'accommodation'],
+        minZoom: 16,
+        layer_ids: [
+            'poi-all',
         ],
         click_layer_ids: 'SAME_AS_LAYER_IDS',
     },