+++ src/app/shared/ojp-sdk/request/location-information/location-information-request-params.interface.ts
@@ -3,8 +3,8 @@ import { GeoRestrictionType, GeoRestrictionPoiOSMTag } from "../../types/geo-res
 export interface LocationInformationRequestParams {
   locationName: string
   stopPlaceRef: string
-  geoRestrictionType: GeoRestrictionType
-  poiOsmTag: GeoRestrictionPoiOSMTag | null
+  geoRestrictionType: GeoRestrictionType | null
+  poiOsmTags: GeoRestrictionPoiOSMTag[] | null
   numberOfResults: number
   bboxWest: number
   bboxNorth: number